static: move stuff in plotConfig.js
