+++ internal/static/src/plot.js
@@ -1,202 +1,14 @@
 import * as theme from "./theme.js";
-import * as ui from "./ui.js";
+import { newConfigObject, newLayoutObject, themeColors } from "./plotConfig.js";
+import { formatFunction } from "./utils.js";
+import { plotWidth, plotHeight } from "./plotConfig.js";
 import Plotly from "plotly.js-cartesian-dist-min";
 import tippy, { followCursor } from "tippy.js";
 import "tippy.js/dist/tippy.css";
 
-const debugMode = true;
-
-const newConfigObject = (cfg, isMaximized) => {
-  return {
-    showEditInChartStudio: debugMode,
-    plotlyServerURL: debugMode ? "https://chart-studio.plotly.com" : ":",
-    displaylogo: false,
-    displayModeBar: true,
-    modeBarButtonsToRemove: [
-      "2D",
-      "zoom2d",
-      "pan2d",
-      "select2d",
-      "lasso2d",
-      "zoomIn2d",
-      "zoomOut2d",
-      "autoScale2d",
-      "resetScale2d",
-      "toggleSpikelines",
-    ],
-    modeBarButtonsToAdd: [
-      {
-        name: "info",
-        title: "Plot info",
-        icon: Plotly.Icons.question,
-        click: ui.onClickPlotInfo,
-      },
-      {
-        name: isMaximized ? "minimize" : "maximize",
-        icon: Plotly.Icons.zoombox,
-        click: ui.onClickPlotMaximize(cfg),
-      },
-    ],
-    toImageButtonOptions: {
-      format: "png",
-      filename: cfg.name,
-      scale: 2,
-    },
-  };
-};
-
-const copyArrayOrNull = (o) => {
-  return (Array.isArray(o) && [...o]) || null;
-};
-
-const plotWidth = 630;
-const plotHeight = 450;
-
-const newLayoutObject = (cfg, isMaximized) => {
-  const layout = {
-    title: {
-      y: 0.94,
-      font: {
-        family: "Roboto",
-        size: 18,
-      },
-      text: cfg.title,
-    },
-    margin: {
-      t: 50,
-      r: 20,
-      l: 60,
-      b: cfg.type === "heatmap" ? 66 : 0,
-    },
-    paper_bgcolor: cfg.layout.paper_bgcolor,
-    plot_bgcolor: cfg.layout.plot_bgcolor,
-    font: {
-      color: cfg.layout.font_color,
-    },
-    width: isMaximized ? null : plotWidth,
-    height: isMaximized ? null : plotHeight,
-    hovermode: "x",
-    barmode: cfg.layout.barmode,
-    xaxis: {
-      tickformat: "%H:%M'%S″",
-      type: "date",
-      fixedrange: true,
-      autorange: false,
-    },
-    yaxis: {
-      tickmode: cfg.layout.yaxis.tickmode,
-      ticktext: copyArrayOrNull(cfg.layout.yaxis.ticktext),
-      tickvals: copyArrayOrNull(cfg.layout.yaxis.tickvals),
-      title: { text: cfg.layout.yaxis.title },
-      fixedrange: true,
-      rangemode: "normal",
-      tickformat: "s",
-      ticksuffix: cfg.layout.yaxis.ticksuffix,
-      minexponent: 3,
-      showexponent: "all",
-      exponentformat: "SI",
-      showticksuffix: "yes",
-      separatethousands: false,
-    },
-    showlegend: true,
-    legend: {
-      orientation: "h",
-      xanchor: "center",
-      x: 0.5,
-      y: -0.05,
-    },
-  };
-
-  if (layout.yaxis.tickmode == "array") {
-    // Format yaxis ticks
-    const formatYUnit = formatFunction(cfg.hover.yunit);
-    for (let i = 0; i < layout.yaxis.ticktext.length; i++) {
-      layout.yaxis.ticktext[i] = formatYUnit(layout.yaxis.ticktext[i]);
-    }
-  }
-
-  return layout;
-};
-
-const themeColors = {
-  light: {
-    paper_bgcolor: "#f8f8f8",
-    plot_bgcolor: "#ffffdd",
-    font_color: "#434343",
-  },
-  dark: {
-    paper_bgcolor: "#181a1c",
-    plot_bgcolor: "#282a2c",
-    font_color: "#fff",
-  },
-};
-
 const plotsDiv = document.getElementById("plots");
 
-/*
-    Plot configuration object:
-    {
-      "name": string,                  // internal name
-      "title": string,                 // plot title 
-      "type": 'scatter'|'bar'|'heatmap' 
-      "updateFreq": int,               // datapoints to receive before redrawing the plot. (default: 1)
-      "infoText": string,              // text showed in the plot 'info' tooltip
-      "events": "lastgc",              // source of vertical lines (example: 'lastgc')
-      "layout": object,                // (depends on plot type)
-      "subplots": array,               // describe 'traces', only for 'scatter' or 'bar' plots
-      "heatmap": object,               // heatmap details
-     }
-
-    Layout for 'scatter' and 'bar' plots:
-    {
-        "yaxis": {
-            "title": {
-                "text": "bytes"      // yaxis title
-            },
-            "ticksuffix": "B",       // base unit for ticks
-        },
-        "barmode": "stack",           // 'stack' or 'group' (only for bar plots)
-    },
-
-    Layout" for heatmaps:
-    {
-        "yaxis": {
-            tickmode:  string  (supports 'array' only)
-            tickvals:  []float64
-            ticktext:  []float64
-            "title": {
-                "text": "size class"
-            }
-    }
-
-    Subplots show the potentially multiple trace objects for 'scatter' and 'bar'
-    plots. Each trace is an object:
-    {
-        "name": string;          // internal name
-        "unitfmt": string,       // d3 format string for tooltip
-        "stackgroup": string,    // stackgroup (if stacked line any)
-        "hoveron": string        // useful for stacked only (TODO(arl): remove from go)
-        "color": colorstring     // plot/trace color
-    }
-
-    Heatmap details object
-    {
-         "colorscale": array      // array of weighted colors,
-         "buckets": array
-         "hover": {
-             "yname": string,     // y axis units
-             "yunit": "bytes",    // y axis name
-             "zname": "objects"   // z axis name 
-         }
-     }
-*/
-
 class Plot {
-  /**
-   * Construct a new Plot object, wrapping a Plotly chart. See above
-   * documentation for plot configuration.
-   */
-
   constructor(cfg) {
     cfg.layout.paper_bgcolor = themeColors[theme.getThemeMode()].paper_bgcolor;
     cfg.layout.plot_bgcolor = themeColors[theme.getThemeMode()].plot_bgcolor;
@@ -457,45 +269,3 @@ const createVerticalLines = (tss) => {
 };
 
 export { createVerticalLines, Plot };
-
-const durUnits = ["w", "d", "h", "m", "s", "ms", "µs", "ns"];
-const durVals = [6048e11, 864e11, 36e11, 6e10, 1e9, 1e6, 1e3, 1];
-
-// Formats a time duration provided in second.
-const formatDuration = (sec) => {
-  let ns = sec * 1e9;
-  for (let i = 0; i < durUnits.length; i++) {
-    let inc = ns / durVals[i];
-
-    if (inc < 1) continue;
-    return Math.round(inc) + durUnits[i];
-  }
-  return res.trim();
-};
-
-const bytesUnits = ["B", "KB", "MB", "GB", "TB", "PB", "EB"];
-
-// Formats a size in bytes.
-const formatBytes = (bytes) => {
-  let i = 0;
-  while (bytes > 1000) {
-    bytes /= 1000;
-    i++;
-  }
-  const res = Math.trunc(bytes);
-  return `${res}${bytesUnits[i]}`;
-};
-
-// Returns a format function based on the provided unit.
-const formatFunction = (unit) => {
-  switch (unit) {
-    case "duration":
-      return formatDuration;
-    case "bytes":
-      return formatBytes;
-  }
-  // Default formatting
-  return (y) => {
-    `${y} ${hover.yunit}`;
-  };
-};
